upgraded phpass-0.1 to 0.3
